Miso.AFrame.Core.Animation

Documentation

data AnimationRepeatCount

Constructors

Finite Int

Indefinite

data AnimationFill

Determines effect of animation when not actively in play.

Constructors

AnimationBackwards

AnimationBoth

AnimationForwards

AnimationNone

data AnimationBasicEasing

Constructors

Ease

EaseIn

EaseOut

EaseInOut

data AnimationEasingGroup

Constructors

Cubic

Quad

Quart

Quint

Sine

Expo

Circ

Elastic

Back

Bounce

data AnimationEasing

Constructors

AnimationEasing

Fields

animationEasingBasic :: AnimationBasicEasing

animationEasingGroup :: Maybe AnimationEasingGroup

data AnimationDirection

Constructors

AnimationAlternate

AnimationAlternateReverse

AnimationNormal

AnimationReverse

data AnimationAttrs

Animation attributes.

Constructors

AnimationAttrs

Fields

animationBegin :: Maybe MisoString

Event name to wait on before beginning animation.

animationEnd :: Maybe MisoString

Event name to wait on before stopping animation.

animationDelay :: Maybe Milliseconds

Delay (in milliseconds) or event name to wait on before beginning animation.

animationDirection :: Maybe AnimationDirection

Direction of the animation (between animationFrom and animationTo).

animationDur :: Maybe Milliseconds

Duration in (milliseconds) of the animation.

animationEasing :: Maybe AnimationEasing

Easing function of the animation. There are very many to choose from.

animationFill :: Maybe AnimationFill

Determines effect of animation when not actively in play.

animationRepeat :: AnimationRepeatCount

Repeat count or Indefinite.

defaultAnimationAttrs :: AnimationAttrs

class ToJSVal a => CanAnimate a

animation

Arguments

:: CanAnimate a

=> MisoString
Name of an attribute to provide animation for.

-> Maybe a
Starting value. Specify Nothing to use current value.

-> a
Ending value.

-> AnimationAttrs
Animation attributes.

-> View action

A-Frame animation entity.